+++ src/runner.rs
@@ -15,6 +15,7 @@ use std::marker::PhantomData;
 use native_tls::{TlsAcceptor, TlsStream, TlsConnector, HandshakeError};
 use dns::{dns_parse};
 use poller::{self, poller};
+use iovec::IoVec;
 
 /// Runner executes fibers on main stack. There can be multiple pollers each with its own fibers
 /// of a particular type.
@@ -76,21 +77,21 @@ impl<P,R> Runner<P,R> {
     ///
     /// This function does not block and fiber gets executed on next poll().
     pub fn new_tcp(&self, tcp: TcpStream, func: FiberFn<P,R>, param: P) -> io::Result<()> {
-        runner(self.pos).register(Some(func), FiberSock::Tcp(tcp), Some(param), None).map(|_|{()})
+        runner(self.pos).register(Some(func), FiberSock::Tcp(tcp), Some(param), None, true).map(|_|{()})
     }
 
     /// Start fiber on TCP listener.
     ///
     /// This function does not block and fiber gets executed on next poll().
     pub fn new_listener(&self, tcp: TcpListener, func: FiberFn<P,R>, param: P) -> io::Result<()> {
-        runner(self.pos).register(Some(func), FiberSock::Listener(tcp), Some(param), None).map(|_|{()})
+        runner(self.pos).register(Some(func), FiberSock::Listener(tcp), Some(param), None, true).map(|_|{()})
     }
 
     /// Start fiber on UDP socket.
     ///
     /// This function does not block and fiber gets executed on next poll().
     pub fn new_udp(&self, udp: UdpSocket, func: FiberFn<P,R>, param: P) -> io::Result<()> {
-        runner(self.pos).register(Some(func), FiberSock::Udp(udp), Some(param), None).map(|_|{()})
+        runner(self.pos).register(Some(func), FiberSock::Udp(udp), Some(param), None, true).map(|_|{()})
     }
 
     /// Return resp after dur. Timer is one-shot.
@@ -101,6 +102,14 @@ impl<P,R> Runner<P,R> {
     /// Run socket-less fiber after dur.
     pub fn new_fiber_timer(&self, dur: Duration, func: FiberFn<P,R>, param: P) -> Option<TimerRef<P,R>> {
         runner::<P,R>(self.pos).new_timer(Some(func), dur, Some(param), None)
+    }
+
+    /// Start a processing fiber. Not tied to a socket or timer. Due to easyfibers being cooperative multitasking
+    /// this fiber will block entire thread until it either finishes executing FiberFn or it calls join_main.
+    ///
+    /// This function does not block and fiber gets executed on next poll().
+    pub fn new_processor(&self, func: FiberFn<P,R>, param: P) -> io::Result<()> {
+        runner::<P,R>(self.pos).register(Some(func), FiberSock::Empty, Some(param), None, true).map(|_|{()})
     }
 }
 
@@ -119,6 +128,8 @@ pub(crate) struct RunnerInt<P,R> {
     free_fibers: VecDeque<usize>,
     tomain_fibers: VecDeque<FiberRef<P,R>>,
     results: VecDeque<R>,
+    // nginx uses a 16K tls buffer for write_bufs operations
+    tmp_buf: [u8;16384],
 }
 
 impl<P,R> RunnerInt<P,R> {
@@ -129,11 +140,41 @@ impl<P,R> RunnerInt<P,R> {
             free_fibers: VecDeque::with_capacity(4),
             tomain_fibers: VecDeque::with_capacity(4),
             results: VecDeque::new(),
+            tmp_buf: [0u8;16384],
         })
     }
 
     pub(crate) fn socket_timeout(&mut self, pos: usize, dur: Option<Duration>) {
         self.fibers[pos].timeout = dur;
+    }
+
+    pub(crate) fn read_bufs(&mut self, pos: usize, buf: &mut [&mut IoVec]) -> io::Result<usize> {
+        loop {
+            match self.fibers[pos].sock.read_bufs(buf) {
+                Ok(r) => {
+                    if r == 0 && buf.len() > 0 {
+                        return Err(io::Error::new(io::ErrorKind::NotConnected,"socket closed"));
+                    } else {
+                        return Ok(r);
+                    }
+                } 
+                Err(err) => {
+                    if err.kind() == io::ErrorKind::WouldBlock {
+                        if !self.fibers[pos].ready.is_readable() {
+                            self.fibers[pos].register(&poller().poller(), Ready::readable())?;
+                        }
+                        self.timed_step_out(pos);
+                        if self.fibers[pos].timed_out > 0 {
+                            self.fibers[pos].timed_out = 0;
+                            return Err(io::Error::new(io::ErrorKind::TimedOut,"read timeout"));
+                        }
+                    } else if err.kind() == io::ErrorKind::Interrupted {
+                    } else {
+                        return Err(err)
+                    }
+                }
+            }
+        }
     }
 
     pub(crate) fn read(&mut self, pos: usize, buf: &mut [u8]) -> io::Result<usize> {
@@ -148,8 +189,6 @@ impl<P,R> RunnerInt<P,R> {
                 } 
                 Err(err) => {
                     if err.kind() == io::ErrorKind::WouldBlock {
-                        // self.fibers[pos].blocking_on = Ready::readable();
-                        // let mut rdy = self.fibers[pos].ready;
                         if !self.fibers[pos].ready.is_readable() {
                             self.fibers[pos].register(&poller().poller(), Ready::readable())?;
                         }
@@ -160,7 +199,35 @@ impl<P,R> RunnerInt<P,R> {
                         }
                     } else if err.kind() == io::ErrorKind::Interrupted {
                     } else {
-                        // self.fibers[pos].state = FiberState::Closed;
+                        return Err(err)
+                    }
+                }
+            }
+        }
+    }
+
+    pub(crate) fn write_bufs(&mut self, pos: usize, bufs: &[&IoVec]) -> io::Result<usize> {
+        loop {
+            match self.fibers[pos].sock.write_bufs(&mut self.tmp_buf, bufs) {
+                Ok(r) =>{
+                    if r == 0 && bufs.len() == 0 {
+                        return Err(io::Error::new(io::ErrorKind::NotConnected,"socket closed"));
+                    } else {
+                        return Ok(r);
+                    }
+                }
+                Err(err) => {
+                    if err.kind() == io::ErrorKind::WouldBlock {
+                        if !self.fibers[pos].ready.is_writable() {
+                            self.fibers[pos].register(&poller().poller(), Ready::writable())?;
+                        }
+                        self.timed_step_out(pos);
+                        if self.fibers[pos].timed_out > 0 {
+                            self.fibers[pos].timed_out = 0;
+                            return Err(io::Error::new(io::ErrorKind::TimedOut,"write timeout"));
+                        }
+                    } else if err.kind() == io::ErrorKind::Interrupted {
+                    } else {
                         return Err(err)
                     }
                 }
@@ -294,7 +361,7 @@ impl<P,R> RunnerInt<P,R> {
         param: Option<P>,
         resp: Option<R>) -> Option<TimerRef<P,R>> {
 
-        if let Ok(pos) = self.register(func, FiberSock::Empty, param, None) {
+        if let Ok(pos) = self.register(func, FiberSock::Empty, param, None, false) {
             self.fibers[pos].result = resp;
             if let Some(tk) = poller().timer_reg(self.fiber_id(pos), Some(dur)) {
                 self.fibers[pos].wtoken = Some(tk);
@@ -335,7 +402,7 @@ impl<P,R> RunnerInt<P,R> {
         if domain.len() == 0 {
             return Err(io::Error::new(io::ErrorKind::Other, "empty domain"));
         }
-        let pos = self.register(func, FiberSock::Empty, Some(param), parent)?;
+        let pos = self.register(func, FiberSock::Empty, Some(param), parent, false)?;
         let mut cp = ConnectParam {
             port,
             st,
@@ -377,9 +444,8 @@ impl<P,R> RunnerInt<P,R> {
     }
 
     pub(crate) fn register(&mut self, func: Option<FiberFn<P,R>>, 
-            ft: FiberSock, param: Option<P>, parent: Option<usize>) -> io::Result<usize> {
+            ft: FiberSock, param: Option<P>, parent: Option<usize>, schedule: bool) -> io::Result<usize> {
 
-        let empty = ft.is_empty();
         let mut fiber = FiberInt {
             ready: Ready::empty(),
             runner: self.pos,
@@ -421,7 +487,7 @@ impl<P,R> RunnerInt<P,R> {
         };
         // Only schedule for immediate execution if it has a socket.
         // Timers and DNS lookups don't.
-        if !empty {
+        if schedule {
             self.push_toexec(pos);
         }
         if let Some(parent) = parent {